# Optimize a couple of nested loops in the scanner.