There are two places in scannerc.go where we iterate through the full simple_keys stack that I think we can avoid:
- when determining [whether a key is possible](https://github.com/go-yaml/yaml/blob/v2/scannerc.go#L641-L647)
- in the [staleness check](https://github.com/go-yaml/yaml/blob/v2/scannerc.go#L844-L861).

I've included a couple of test cases that show the bad behavior, even with the recent depth limits :

```
$ git checkout c9e8ea1b90fe07c9447c7cbcc212e927d6988388 && go test -run=NONE -bench=. -benchmem ./... > old.txt
$ git checkout 3717a577a65196747809a63ffe76c5ca6b0d88e8 && go test -run=NONE -bench=. -benchmem ./... > new.txt
$ benchcmp old.txt new.txt 
benchmark                                old ns/op       new ns/op      delta
Benchmark1000KB100Aliases-6              1056925022      1094599057     +3.56%
Benchmark1000KBDeeplyNestedSlices-6      282869888       7835556        -97.23%
Benchmark1000KBDeeplyNestedMaps-6        266289230       7375439        -97.23%
Benchmark1000KBDeeplyNestedIndents-6     4360666         4514662        +3.53%
Benchmark1000KB1000IndentLines-6         449900306       451543002      +0.37%
Benchmark1KBMaps-6                       599228          620644         +3.57%
Benchmark10KBMaps-6                      5801815         5892398        +1.56%
Benchmark100KBMaps-6                     53699570        57544454       +7.16%
Benchmark1000KBMaps-6                    513662800       538354297      +4.81%
BenchmarkDeepSlice-6                     46967064873     448771530      -99.04%
BenchmarkDeepFlow-6                      45121229024     648750899      -98.56%

benchmark                                old allocs     new allocs     delta
Benchmark1000KB100Aliases-6              5832441        5832438        -0.00%
Benchmark1000KBDeeplyNestedSlices-6      9068           9249           +2.00%
Benchmark1000KBDeeplyNestedMaps-6        9072           9252           +1.98%
Benchmark1000KBDeeplyNestedIndents-6     10082          10084          +0.02%
Benchmark1000KB1000IndentLines-6         4093516        4093517        +0.00%
Benchmark1KBMaps-6                       3126           3128           +0.06%
Benchmark10KBMaps-6                      30780          30782          +0.01%
Benchmark100KBMaps-6                     307269         307271         +0.00%
Benchmark1000KBMaps-6                    3072079        3072082        +0.00%
BenchmarkDeepSlice-6                     2048121        2048304        +0.01%
BenchmarkDeepFlow-6                      3334112        3334113        +0.00%

benchmark                                old bytes     new bytes     delta
Benchmark1000KB100Aliases-6              393120640     393118752     -0.00%
Benchmark1000KBDeeplyNestedSlices-6      4689806       4748760       +1.26%
Benchmark1000KBDeeplyNestedMaps-6        4690070       4749212       +1.26%
Benchmark1000KBDeeplyNestedIndents-6     2969933       2970002       +0.00%
Benchmark1000KB1000IndentLines-6         143718544     143718560     +0.00%
Benchmark1KBMaps-6                       218984        219112        +0.06%
Benchmark10KBMaps-6                      2178160       2178284       +0.01%
Benchmark100KBMaps-6                     22002813      22002938      +0.00%
Benchmark1000KBMaps-6                    220560496     220560768     +0.00%
BenchmarkDeepSlice-6                     120114824     120174234     +0.05%
BenchmarkDeepFlow-6                      190656952     190656984     +0.00%

```