@liggitt 2019-11-05
v2 is the default branch. v3 is in prerelease.