@niemeyer 2019-11-05
Let's get the issue fixed in v2 first and then we can easily port it forward to v3.

I will have a deeper look into the issue and the suggested solution and get back to you soon.