@cjcullen 2019-11-06
First attempt had a bug dealing w/ complex keys added to the stack immediately after another key was "removed" (keys are not actually taken out of the stack, but do get overwritten, and the new key might not be a possible simple_key, so min_possible needs to account for that).

```
$ git checkout c292eee44f0f38e8507803c26868dc8b889e2550 && go test -run=NONE -bench=. -benchmem ./... > old.txt
$ git checkout 01c0521edf7a0d83650270898a38972a8d8264b4 && go test -run=NONE -bench=. -benchmem ./... > new.txt
$ benchcmp old.txt new.txt
benchmark                                old ns/op       new ns/op     delta
Benchmark1000KB100Aliases-6              933118408       924264884     -0.95%
Benchmark1000KBDeeplyNestedSlices-6      292079720       8114551       -97.22%
Benchmark1000KBDeeplyNestedMaps-6        276815596       8199085       -97.04%
Benchmark1000KBDeeplyNestedIndents-6     4605536         4406578       -4.32%
Benchmark1000KB1000IndentLines-6         446018719       447012500     +0.22%
Benchmark1KBMaps-6                       606299          597170        -1.51%
Benchmark10KBMaps-6                      5848238         6323868       +8.13%
Benchmark100KBMaps-6                     55045361        57967332      +5.31%
Benchmark1000KBMaps-6                    502323739       533391064     +6.18%
BenchmarkDeepSlice-6                     46958325534     439644015     -99.06%
BenchmarkDeepFlow-6                      44440589216     601371522     -98.65%

benchmark                                old allocs     new allocs     delta
Benchmark1000KB100Aliases-6              5832424        5832427        +0.00%
Benchmark1000KBDeeplyNestedSlices-6      9066           9248           +2.01%
Benchmark1000KBDeeplyNestedMaps-6        9072           9251           +1.97%
Benchmark1000KBDeeplyNestedIndents-6     10083          10084          +0.01%
Benchmark1000KB1000IndentLines-6         4093516        4093517        +0.00%
Benchmark1KBMaps-6                       3126           3128           +0.06%
Benchmark10KBMaps-6                      30780          30782          +0.01%
Benchmark100KBMaps-6                     307269         307271         +0.00%
Benchmark1000KBMaps-6                    3072079        3072081        +0.00%
BenchmarkDeepSlice-6                     2048122        2048307        +0.01%
BenchmarkDeepFlow-6                      3334111        3334113        +0.00%

benchmark                                old bytes     new bytes     delta
Benchmark1000KB100Aliases-6              393118196     393119108     +0.00%
Benchmark1000KBDeeplyNestedSlices-6      4689654       4748743       +1.26%
Benchmark1000KBDeeplyNestedMaps-6        4690070       4748824       +1.25%
Benchmark1000KBDeeplyNestedIndents-6     2969955       2969997       +0.00%
Benchmark1000KB1000IndentLines-6         143718512     143718560     +0.00%
Benchmark1KBMaps-6                       218984        219112        +0.06%
Benchmark10KBMaps-6                      2178160       2178281       +0.01%
Benchmark100KBMaps-6                     22002906      22002931      +0.00%
Benchmark1000KBMaps-6                    220560496     220560672     +0.00%
BenchmarkDeepSlice-6                     120114936     120173917     +0.05%
BenchmarkDeepFlow-6                      190656856     190656976     +0.00%
```