@niemeyer 2019-11-06
It's good to avoid force pushing in those cases, so that we can see what changed between the revisions.

Either way, as we chatted over email I hadn't touched this logic from the ported C library so far as it seemed to do its job, but I'd like to try taking a step back before adding more complexity to see if there's an alternative approach that could get rid of the issues found while simplifying the implementation instead of making it harder to understand and maintain.

I believe this is the case here. I will push an alternative algorithm to v3 that is both simpler and faster so you can have a look.