@cjcullen 2019-11-07
TL;DR: I think your patch fixes the issue, but I got fairly confused along the way :).
Do you want me to open a PR for your patch on v2? And maybe a separate one to add the two testcases that it fixes?
...

I haven't looked to much at the code on v3, but it appears to be more strict on some sketchy (but I think valid) YAML. E.g.
```
{a:{}}
```
produces: `unexpected error: yaml: did not find expected ',' or '}'`
And
```
{a,a}
```
produces: `line 1: mapping key "a" already defined at line 1` (I think because `d.uniqueKeys` defaults to true now?)


The DeepFlow benchmark I added would now fail because of the reasons above, but runs out of memory on my machine before it can fail. That test looks something like:
```
{a,b:
 {a,b:{a,b:{a,b:{1,1,1,1,1,1,1,1}}}}}
```
Where the maps are nested to max depth, with ~1000kB of crap (~500,000 tokens) at the bottom.

I made the bottom map in the test smaller to gather some numbers.
```
100 tokens:
BenchmarkDeepFlow-6   	      99	  12571431 ns/op	 7661837 B/op	   40271 allocs/op
1000 tokens:
BenchmarkDeepFlow-6   	       5	 218142322 ns/op	101838585 B/op	 2022118 allocs/op
10000 tokens:
BenchmarkDeepFlow-6   	       1	18689774500 ns/op	9762953496 B/op	200040215 allocs/op
```

I think this is because we are generating per-pair error messages for each re-defined key here: https://github.com/go-yaml/yaml/blob/v3/decode.go#L749-L760. That is probably fairly easy to clean up.

When I flip `d.uniqueKeys` to false, I still see a notable performance degradation from the fixed v2 to the fixed v3
```
$ git checkout cjv3 && go test -run=NONE -bench=. -benchmem ./... > v3.txt
$ git checkout deepflow && go test -run=NONE -bench=. -benchmem ./... > v2.txt
$ benchcmp v2.txt v3.txt
benchmark                                old ns/op     new ns/op      delta
Benchmark1000KB100Aliases-6              922020576     1720174932     +86.57%
Benchmark1000KBDeeplyNestedSlices-6      8025403       60725443       +656.67%
Benchmark1000KBDeeplyNestedMaps-6        8060659       59798584       +641.86%
Benchmark1000KBDeeplyNestedIndents-6     4701986       15094606       +221.03%
Benchmark1000KB1000IndentLines-6         451119036     808144598      +79.14%
Benchmark1KBMaps-6                       630891        1199927        +90.20%
Benchmark10KBMaps-6                      6462013       15317658       +137.04%
Benchmark100KBMaps-6                     54991397      124880821      +127.09%
Benchmark1000KBMaps-6                    516057186     1114736076     +116.01%
BenchmarkDeepSlice-6                     439363971     1091913514     +148.52%
BenchmarkDeepFlow-6                      614039880     1259611292     +105.14%

benchmark                                old allocs     new allocs     delta
Benchmark1000KB100Aliases-6              5832426        6344290        +8.78%
Benchmark1000KBDeeplyNestedSlices-6      9247           9090           -1.70%
Benchmark1000KBDeeplyNestedMaps-6        9252           9095           -1.70%
Benchmark1000KBDeeplyNestedIndents-6     10084          10107          +0.23%
Benchmark1000KB1000IndentLines-6         4093517        4093555        +0.00%
Benchmark1KBMaps-6                       3128           3651           +16.72%
Benchmark10KBMaps-6                      30782          35923          +16.70%
Benchmark100KBMaps-6                     307271         358502         +16.67%
Benchmark1000KBMaps-6                    3072081        3584125        +16.67%
BenchmarkDeepSlice-6                     2048293        2540164        +24.01%
BenchmarkDeepFlow-6                      3334113        3816152        +14.46%

benchmark                                old bytes     new bytes      delta
Benchmark1000KB100Aliases-6              393118424     1470212888     +273.99%
Benchmark1000KBDeeplyNestedSlices-6      4748518       13930102       +193.36%
Benchmark1000KBDeeplyNestedMaps-6        4749088       13930332       +193.33%
Benchmark1000KBDeeplyNestedIndents-6     2969992       12260117       +312.80%
Benchmark1000KB1000IndentLines-6         143718560     697303904      +385.19%
Benchmark1KBMaps-6                       219112        649889         +196.60%
Benchmark10KBMaps-6                      2178283       11417515       +424.15%
Benchmark100KBMaps-6                     22002941      114370482      +419.80%
Benchmark1000KBMaps-6                    220560624     1297664104     +488.35%
BenchmarkDeepSlice-6                     120172712     1182242296     +883.79%
BenchmarkDeepFlow-6                      190657024     1067076568     +459.68%
```

I think most of that is just v3-specific stuff (I don't think I'm running into [this](https://github.com/go-yaml/yaml/issues/520#issuecomment-540647966), but I could be), because comparing it to v2 w/ just your patch applied gets much better numbers:
```
$ benchcmp v2_cj.txt v2_gustavo.txt
benchmark                                old ns/op     new ns/op     delta
Benchmark1000KB100Aliases-6              922020576     928285149     +0.68%
Benchmark1000KBDeeplyNestedSlices-6      8025403       46426178      +478.49%
Benchmark1000KBDeeplyNestedMaps-6        8060659       45873584      +469.10%
Benchmark1000KBDeeplyNestedIndents-6     4701986       4353733       -7.41%
Benchmark1000KB1000IndentLines-6         451119036     437655646     -2.98%
Benchmark1KBMaps-6                       630891        591242        -6.28%
Benchmark10KBMaps-6                      6462013       5891734       -8.83%
Benchmark100KBMaps-6                     54991397      52778245      -4.02%
Benchmark1000KBMaps-6                    516057186     487526584     -5.53%
BenchmarkDeepSlice-6                     439363971     454282346     +3.40%
BenchmarkDeepFlow-6                      614039880     595380818     -3.04%

benchmark                                old allocs     new allocs     delta
Benchmark1000KB100Aliases-6              5832426        5832427        +0.00%
Benchmark1000KBDeeplyNestedSlices-6      9247           9066           -1.96%
Benchmark1000KBDeeplyNestedMaps-6        9252           9070           -1.97%
Benchmark1000KBDeeplyNestedIndents-6     10084          10082          -0.02%
Benchmark1000KB1000IndentLines-6         4093517        4093516        -0.00%
Benchmark1KBMaps-6                       3128           3126           -0.06%
Benchmark10KBMaps-6                      30782          30780          -0.01%
Benchmark100KBMaps-6                     307271         307269         -0.00%
Benchmark1000KBMaps-6                    3072081        3072079        -0.00%
BenchmarkDeepSlice-6                     2048293        2048121        -0.01%
BenchmarkDeepFlow-6                      3334113        3334111        -0.00%

benchmark                                old bytes     new bytes     delta
Benchmark1000KB100Aliases-6              393118424     393119352     +0.00%
Benchmark1000KBDeeplyNestedSlices-6      4748518       4689528       -1.24%
Benchmark1000KBDeeplyNestedMaps-6        4749088       4689814       -1.25%
Benchmark1000KBDeeplyNestedIndents-6     2969992       2969939       -0.00%
Benchmark1000KB1000IndentLines-6         143718560     143718512     -0.00%
Benchmark1KBMaps-6                       219112        218984        -0.06%
Benchmark10KBMaps-6                      2178283       2178155       -0.01%
Benchmark100KBMaps-6                     22002941      22002804      -0.00%
Benchmark1000KBMaps-6                    220560624     220560496     -0.00%
BenchmarkDeepSlice-6                     120172712     120114866     -0.05%
BenchmarkDeepFlow-6                      190657024     190656896     -0.00%
```