@cjcullen 2019-11-09
In my testing, your patch ported onto v2 was not significantly slower, except in a couple of the max-depth error cases, and even then, we're talking about 50ms vs. 1ms.
```
$ benchcmp cj.txt gus.txt
benchmark                                old ns/op      new ns/op      delta
Benchmark1000KB100Aliases-6              1043562949     1050274409     +0.64%
Benchmark1000KBDeeplyNestedSlices-6      8077086        48635551       +502.14%
Benchmark1000KBDeeplyNestedMaps-6        8580521        50397244       +487.34%
Benchmark1000KBDeeplyNestedIndents-6     4508057        4559411        +1.14%
Benchmark1000KB1000IndentLines-6         455249763      436821141      -4.05%
Benchmark1KBMaps-6                       626294         593838         -5.18%
Benchmark10KBMaps-6                      5938881        6193108        +4.28%
Benchmark100KBMaps-6                     55397757       54890679       -0.92%
Benchmark1000KBMaps-6                    532778194      493524839      -7.37%
BenchmarkDeepSlice-6                     448975265      456474390      +1.67%
BenchmarkDeepFlow-6                      631368987      403562717      -36.08%

benchmark                                old allocs     new allocs     delta
Benchmark1000KB100Aliases-6              5832440        5832440        +0.00%
Benchmark1000KBDeeplyNestedSlices-6      9249           9066           -1.98%
Benchmark1000KBDeeplyNestedMaps-6        9252           9071           -1.96%
Benchmark1000KBDeeplyNestedIndents-6     10083          10083          +0.00%
Benchmark1000KB1000IndentLines-6         4093517        4093516        -0.00%
Benchmark1KBMaps-6                       3128           3126           -0.06%
Benchmark10KBMaps-6                      30782          30780          -0.01%
Benchmark100KBMaps-6                     307271         307269         -0.00%
Benchmark1000KBMaps-6                    3072081        3072079        -0.00%
BenchmarkDeepSlice-6                     2048297        2048121        -0.01%
BenchmarkDeepFlow-6                      3334113        1978105        -40.67%

benchmark                                old bytes     new bytes     delta
Benchmark1000KB100Aliases-6              393119144     393120544     +0.00%
Benchmark1000KBDeeplyNestedSlices-6      4748894       4689568       -1.25%
Benchmark1000KBDeeplyNestedMaps-6        4748926       4689851       -1.24%
Benchmark1000KBDeeplyNestedIndents-6     2969957       2969951       -0.00%
Benchmark1000KB1000IndentLines-6         143718592     143718512     -0.00%
Benchmark1KBMaps-6                       219112        218984        -0.06%
Benchmark10KBMaps-6                      2178284       2178156       -0.01%
Benchmark100KBMaps-6                     22002935      22002808      -0.00%
Benchmark1000KBMaps-6                    220560624     220560496     -0.00%
BenchmarkDeepSlice-6                     120173610     120114824     -0.05%
BenchmarkDeepFlow-6                      190656976     115058064     -39.65%
```